Move tests to welcome folder
