+++ test/widgets/welcome/after_registration_page_test.dart
@@ -5,7 +5,7 @@ import 'package:motis_mitfahr_app/welcome/pages/after_registration_page.dart';
 import 'package:motis_mitfahr_app/welcome/pages/login_page.dart';
 import 'package:motis_mitfahr_app/welcome/pages/welcome_page.dart';
 
-import '../util/pump_material.dart';
+import '../../util/pump_material.dart';
 
 void main() {
   setUpAll(() async {
