+++ test/widgets/welcome/forgot_password_page_test.dart
@@ -6,11 +6,11 @@ import 'package:motis_mitfahr_app/util/supabase.dart';
 import 'package:motis_mitfahr_app/welcome/pages/forgot_password_page.dart';
 import 'package:progress_state_button/progress_button.dart';
 
-import '../util/mocks/mock_server.dart';
-import '../util/mocks/navigator_observer.mocks.dart';
-import '../util/mocks/request_processor.dart';
-import '../util/mocks/request_processor.mocks.dart';
-import '../util/pump_material.dart';
+import '../../util/mocks/mock_server.dart';
+import '../../util/mocks/navigator_observer.mocks.dart';
+import '../../util/mocks/request_processor.dart';
+import '../../util/mocks/request_processor.mocks.dart';
+import '../../util/pump_material.dart';
 
 void main() {
   final MockRequestProcessor processor = MockRequestProcessor();
