+++ test/widgets/welcome/register_page_test.dart
@@ -7,10 +7,10 @@ import 'package:motis_mitfahr_app/welcome/pages/after_registration_page.dart';
 import 'package:motis_mitfahr_app/welcome/pages/register_page.dart';
 import 'package:progress_state_button/progress_button.dart';
 
-import '../util/mocks/mock_server.dart';
-import '../util/mocks/request_processor.dart';
-import '../util/mocks/request_processor.mocks.dart';
-import '../util/pump_material.dart';
+import '../../util/mocks/mock_server.dart';
+import '../../util/mocks/request_processor.dart';
+import '../../util/mocks/request_processor.mocks.dart';
+import '../../util/pump_material.dart';
 
 void main() {
   final MockRequestProcessor processor = MockRequestProcessor();
